Fix / clean up summary ops (separate for train and valid)
Key things to note:
1) Two separate FileWriters. Second one (for valid set) doesn't
pass in graph because it's the same graph
2) When running training, run summary_op from tf.summary.merge_all()
- This runs all of the summary ops
3) When running validation, just run the loss summary and acc summary ops, but
now with validation data passed in

Note: training step vs valid set is so much slower... and it's not because of the
backpro step... it's a direct function of the size. Might be a problem with feed_dict.
Not sure if it's supposed to be used if I'm using queues?

Test plan:
1) python tasks/image-sent/main.py --arch basic_cnn -obj sent_triclass --mode train --val_every_epoch 1
2) Go to tensorboard
- See plots on same graph