Speedup by putting imgs into graph for train set
Instead of using feed_dict (which is slower) for both train and valid,
only do it for valid. Do this by using tf.placeholder_with_default instead
of tf.placeholder, where the default is the train_img_batch passed into
the model when setting up the graph. If a feeddict is passed in to overwrite
that input tensor, then use it. So pass in feeddict when using valid

Speedup: Maybe ~2.1 seconds per minibatch of size 16 on training to
~1.8 seconds

Test plan:
1) python tasks/image-sent/main.py --arch basic_cnn -obj sent_triclass --mode train
- Faster, valid still runs

Note: valid still way faster. Again, this is because train queue is so much bigger.
Need to shard the train queue somehow? Probably going to train tfrecords next