[Website] More efficient server-js comm; Fix bug w/ clearing chart
1) Instead of passing a title2framepaths to js, just pass
the framepaths for the current video. I feel like when we have tens
or hundreds of videos this data structure could hurt performance.
Granted, now we have to pass framepaths everytime we change a dat.gui
param and hit /api/pred/title/window_len, but I don't think that's so bad.

2) There was a bug where sometimes on hover would glitch out if the
video had been changed a number of times. This is because it has to be cleared.
So have global var for the lineChart. And call destroy everytime it's redrawn.